llm-backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.34.1
anthropic-bedrock==0.8.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
asyncio==3.4.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
biopython==1.83
biothings-client==0.3.1
blinker==1.7.0
boto3==1.34.122
botocore==1.34.122
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
cattrs==23.2.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.5
dataclasses-json==0.6.4
datasets==2.18.0
dateparser==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
easyocr==1.7.1
etelemetry==0.3.1
executing==2.0.1
fastapi==0.110.1
filelock==3.13.4
fitz==0.0.1.dev2
Flask==3.0.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
gprofiler-official==1.0.0
gpt4all==2.7.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmldate==1.8.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
humanize==4.9.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
instructor==1.3.7
ipython==8.23.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.7
kubernetes==29.0.0
langchain==0.2.14
langchain-anthropic==0.1.23
langchain-aws==0.1.6
langchain-community==0.2.12
langchain-core==0.2.33
langchain-experimental==0.0.64
langchain-mongodb==0.1.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langgraph==0.2.4
langgraph-checkpoint==1.0.3
langsmith==0.1.75
lazy_loader==0.4
llama-index==0.10.43
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.43
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-chroma==0.1.9
llama-index-vector-stores-qdrant==0.2.9
llama-parse==0.4.4
llama_cpp_python==0.2.77
llamaindex-py-client==0.1.19
looseversion==1.3.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==1.0.16
mmh3==4.1.0
monotonic==1.6
motor==3.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mygene==3.2.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.8
onnxruntime==1.18.0
open-clip-torch==2.24.0
openai==1.30.1
opencv-python-headless==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
portalocker==2.8.2
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
prov==2.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==2.0.0
Pygments==2.17.2
PyJWT==2.8.0
pymed==0.8.9
pymongo==4.7.1
PyMuPDF==1.24.12
pymupdf4llm==0.0.17
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pyphen==0.15.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pysbd==0.3.4
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-pptx==0.6.23
python-socketio==5.11.2
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qdrant-client==1.8.2
ragas==0.1.7
rank-bm25==0.2.2
rdflib==7.0.0
redis==5.0.3
regex==2024.4.16
requests==2.31.0
requests-cache==1.2.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.21.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
tenacity>=8.4.1,<9.0.0
textstat==0.7.3
threadpoolctl==3.4.0
tifffile==2024.5.22
tiktoken==0.7.0
timm==0.9.16
tokenizers==0.15.2
tools==0.1.9
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
traits==6.3.2
transformers==4.39.3
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
url-normalize==1.4.3
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
